Add files via upload
 Welcome to the Game Of Life program!
*****************************************************
It is a zero-player game, meaning that its evolution
is determined by its initial state, requiring no further
input. One interacts with the Game of Life by creating an
initial configuration and observing how it evolves.

Rule - The universe of the Game of Life is an infinite,
two-dimensional orthogonal grid of square cells,
each of which is in one of two possible states, live or dead.
Every cell interacts with its eight neighbours.